Tile is now an abstract class. Pnts is its implementation for point tiles
